[Kernel] Modernize CDNA SageAttention for MI308X (gfx942)
Split sage_attn_cdna into pipeline + sage_attn_helpers/utils; migrate K/Q/O
GM loads to make_buffer_tensor + copy_atom; fx vector ops and @flyc.jit mask
dispatch; gate tests to MI308X; cleanup gfx950 dead code and formatting.
Descale/bias GM remains on pointer load (bit-exact on 0721-zimage).

Validated: harness S256/S4608/S8448 sha256 bit-exact; test_sage_attn pass.
